+++ lib/tzispa/rig.rb
@@ -1,11 +1,12 @@
 module Tzispa
   module Rig
-
+    
+    require 'tzispa/rig/version'
     require 'tzispa/rig/parameters'
+    require 'tzispa/rig/parsernext'
     require 'tzispa/rig/binder'
     require 'tzispa/rig/template'
-    require 'tzispa/rig/parsernext'
-    require 'tzispa/rig/version'
+    require 'tzispa/rig/engine'
 
   end
 end